pyproject.toml
[tool.poetry]
name = "repl-nix-chuck-noris-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_chuck_noris_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.4.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
